Create bare bones of pong game
Create board class which handles components displayed
on the board. This includes two paddles, and the
center divide line. Correctly positions the paddles
at opposite ends of the board.

Create paddle class which represents the paddle. It
is able to move up and down until it touches the top
wall. Collision detection with bottom still needs to
be implemented.